+++ src/hapi/pipelines/utilities/parse_tags.py
@@ -1,13 +1,17 @@
+import re
+
+from hapi_schema.utils.enums import Gender
 from hdx.utilities.text import multiple_replace
 from hxl.model import Column, TagPattern
 
 
 def get_gender_and_age_range(hxl_tag: str) -> (str, str):
-    gender = "*"
-    age_range = "*"
+    gender = "all"
+    age_range = "all"
     col = Column.parse(hxl_tag)
     gender_patterns = {
-        TagPattern.parse(f"#*+{g}"): g for g in ["f", "m", "x", "u", "o", "e"]
+        TagPattern.parse(f"#*+{gender.value}"): gender.value
+        for gender in Gender
     }
     for pattern in gender_patterns:
         if pattern.match(col):
@@ -23,18 +27,20 @@ def get_gender_and_age_range(hxl_tag: str) -> (str, str):
     age_component = multiple_replace(
         age_component[0], {"age_": "", "age": "", "_age": ""}
     )
-    if age_component.endswith("plus"):
-        age_range = multiple_replace(
-            age_component, {"plus": "+", "_plus": "+"}
-        )
-    else:
-        age_range = age_component.replace("_", "-")
+    # Check if age component contains any numbers
+    if re.search(r"\d", age_component):
+        if age_component.endswith("plus"):
+            age_range = multiple_replace(
+                age_component, {"plus": "+", "_plus": "+"}
+            )
+        else:
+            age_range = age_component.replace("_", "-")
 
     return gender, age_range
 
 
 def get_min_and_max_age(age_range: str) -> (int | None, int | None):
-    if age_range == "*" or age_range == "unknown":
+    if age_range == "all" or age_range == "unknown":
         return None, None
     ages = age_range.split("-")
     if len(ages) == 2: