+++ tests/test_main.py
@@ -178,7 +178,7 @@ class TestHAPIPipelines:
                     count = session.scalar(
                         select(func.count(DBRefugees.resource_hdx_id))
                     )
-                    check.equal(count, 65355)
+                    check.equal(count, 56641)
                     count = session.scalar(
                         select(func.count(DBFunding.resource_hdx_id))
                     )